Улучшение пагинации и добавление лоадера
