+++ src/config/apiRequests.tsx
@@ -17,22 +17,22 @@ type ProductProps = {
   description: string;
 };
 
-const token =
-  'f53a84efed5478ffc79d455646b865298d6531cf8428a5e3157fa5572c6d3c51739cdaf3a28a4fdf8b83231163075ef6a8435a774867d035af53717fecd37bca814c6b7938f02d2893643e2c1b6a2f79b3ca715222895e8ee9374c0403d44081e135cda1f811fe7cfec6454746a5657ba070ec8456462f8ca0e881232335d1ef';
+// const token =
+//   'f53a84efed5478ffc79d455646b865298d6531cf8428a5e3157fa5572c6d3c51739cdaf3a28a4fdf8b83231163075ef6a8435a774867d035af53717fecd37bca814c6b7938f02d2893643e2c1b6a2f79b3ca715222895e8ee9374c0403d44081e135cda1f811fe7cfec6454746a5657ba070ec8456462f8ca0e881232335d1ef';
 
-export const fetchProducts = async () => {
-  const query = qs.stringify({
-    populate: ['images', 'productCategory'],
-  });
+// export const fetchProducts = async () => {
+//   const query = qs.stringify({
+//     populate: ['images', 'productCategory'],
+//   });
 
-  const response = await axios.get(`${STRAPI_URL}/products?${query}`, {
-    headers: {
-      Authorization: `Bearer ${token}`,
-    },
-  });
+//   const response = await axios.get(`${STRAPI_URL}/products?${query}`, {
+//     headers: {
+//       Authorization: `Bearer ${token}`,
+//     },
+//   });
 
-  return response.data.data;
-};
+//   return response.data.data;
+// };
 
 export const fetchSingleProduct = async (documentId: string | undefined) => {
   const query = qs.stringify({
