+++ src/pages/products/components/Catalog/Catalog.tsx
@@ -14,7 +14,11 @@ import ArrowRightIcon from 'components/icons/ArrowRightIcon';
 import { Link } from 'react-router';
 import { routes } from 'config/routes';
 
-import { fetchProducts } from 'config/apiRequests';
+import axios from 'axios';
+import qs from 'qs'
+
+//import { fetchProducts } from 'config/apiRequests';
+import Loader from 'components/Loader';
 
 const Catalog = () => {
   //Интерфейс для состояния с товарами
@@ -35,11 +39,32 @@ const Catalog = () => {
 
   const [products, setProducts] = useState<IProducts[]>([]);
   const [page, setPage] = useState<number>(1);
+  const [totalProducts, setTotalProducts] = useState<number>(0)
   const productsPerPage = 12;
 
   useEffect(() => {
-    fetchProducts().then((data) => setProducts(data));
-  }, []);
+    const token =
+    'f53a84efed5478ffc79d455646b865298d6531cf8428a5e3157fa5572c6d3c51739cdaf3a28a4fdf8b83231163075ef6a8435a774867d035af53717fecd37bca814c6b7938f02d2893643e2c1b6a2f79b3ca715222895e8ee9374c0403d44081e135cda1f811fe7cfec6454746a5657ba070ec8456462f8ca0e881232335d1ef';
+  
+  const fetchProducts = async () => {
+    const query = qs.stringify({
+      populate: ['images', 'productCategory'],
+    });
+  
+    const response = await axios.get(`
+https://front-school-strapi.ktsdev.ru/api/products?populate[0]=images&populate[1]=productCategory&pagination[pageSize]=12&pagination[page]=${page}`, {
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    }).then((res) => {setProducts(res.data.data);setTotalProducts(res.data.meta.pagination.total)})
+  };
+
+  fetchProducts();
+  }, [])
+
+  // useEffect(() => {
+  //   fetchProducts().then((data) => setProducts(data));
+  // }, []);
 
   const handleChange = () => {};
 
@@ -73,9 +98,7 @@ const Catalog = () => {
           <Text view="subtitle" weight="bold">
             Total products
           </Text>
-          <Text view="p-20" weight="bold" color="accent">
-            {products.length}
-          </Text>
+          {totalProducts !== 0 ? <Text view='p-20' weight='bold' color='accent'>{totalProducts}</Text> : <Loader size='m' className={styles.loadingRotating}/>}
         </div>
       </div>
       <div className={styles.catalogBody}>
@@ -94,7 +117,7 @@ const Catalog = () => {
           ))}
       </div>
       <div className={styles.pagination}>
-        <ArrowRightIcon
+      <ArrowRightIcon
           width={35}
           height={35}
           color={page === 1 ? 'secondary' : 'primary'}
@@ -113,60 +136,6 @@ const Catalog = () => {
                 }
           }
         />
-        <button
-          onClick={() => {
-            setPage(1);
-            setTimeout(() => {
-              catalogRef.current?.scrollIntoView({
-                block: 'start',
-                behavior: 'smooth',
-              });
-            }, 0);
-          }}
-          className={`${styles.pageSelector} ${page === 1 ? styles.active : ''}`}
-        >
-          1
-        </button>
-        {page > 3 && <span className={styles.ellipsis}>...</span>}
-        {Array.from({ length: totalPages })
-          .map((_, i) => i + 1)
-          .filter((p) => (p >= page - 1 && p <= page + 1) || p === totalPages)
-          .filter((p) => p > 1 && p < totalPages)
-          .map((p) => (
-            <button
-              key={p}
-              onClick={() => {
-                setPage(p);
-                setTimeout(() => {
-                  catalogRef.current?.scrollIntoView({
-                    block: 'start',
-                    behavior: 'smooth',
-                  });
-                }, 0);
-              }}
-              className={`${styles.pageSelector} ${page === p ? styles.active : ''}`}
-            >
-              {p}
-            </button>
-          ))}
-        {page < totalPages - 2 && <span className={styles.ellipsis}>...</span>}
-        {totalPages > 1 && (
-          <button
-            onClick={() => {
-              setPage(totalPages);
-              setTimeout(() => {
-                catalogRef.current?.scrollIntoView({
-                  block: 'start',
-                  behavior: 'smooth',
-                });
-              }, 0);
-            }}
-            className={`${styles.pageSelector} ${page === totalPages ? styles.active : ''}`}
-          >
-            {totalPages}
-          </button>
-        )}
-
         <ArrowRightIcon
           width={35}
           height={35}
